+++ config.toml
@@ -1,5 +1,5 @@
 [model]
-hidden_sizes = [1000, 1000, 1000]
+hidden_sizes = [250, 250, 250]
 epochs = 10000
 loss_threshold = 1
 damping_factor = 0.7
@@ -8,4 +8,4 @@ learning_rate = 0.00005
 skip_profiling = 1
 
 [device]
-device = "cuda"
+device = "mps"
